# optmize proofs contrustion algorithm