@jeroiraz 2020-09-16
Thanks @leogr  for your contribution.
immudb status update protocol compares against latest merkle tree but this optimization may be useful later on by immudb or by any other system as well.